New generic no-edit marking functions
The built-in Ivy marking function depend on editing the completion
itself with `ivy-mark-prefix` so that the marked state gets passed
through all the display transformers and format functions to the end
user.

However, I’ve (tentatively) implemented a custom overriding format
function in this config that will show the marked status of candidates
only based on `ivy-marked-candidates`. These replacement mark
functions do not try to change the completion candidates so they do
not cause errors when the candidate is not a string (as in the case of
`swiper-isearch` candidates).