Bind ivy-immediate-done to M-RET instead
<C-return> is not available on all terminals since some terminals send
<return> as C-j. M-RET, however, is always available, possibly using
ESC RET if necessary.

The old binding for M-RET, `ivy-alt-done`, is discarded because it is
equivalent to ivy-done, except in the case of file selection where it
is used to continue completion from the selected directory instead of
exiting with the selected directory. However, continuing completion
from the current directory can be achieved more naturally anyway by
typing “/” when a directory is selected.